autotools
Address all warnings in autoreconf, which are mainly automake
warnings.

Add SRCDIR sanity check.

Run autoupdate and take all changes except the version bump.

Fix bug that caused odd installation of binary executable.

  $ make install DESTDIR=$dest
  $ cd $dest
  $ find -type f
  ...
  ./usr/local/include/ColPack/Definitions.h
  ./usr/local/include/ColPack/ColPackHeaders.h
  ./home/barak/src/git/colpack/progs/ColPack <---- Wot?

The intent seems to be to build this, but not install it.
Make it so, with tweak to Makefile.am

correct version number in configure.ac