Updated OpenMP directives to comply with the G++ 4.9.2 compiler, which no longer permits the invalid/nonexistent `#pragma omp end critical` directive.
This simultaneously stripped extraneous whitespace at the end of lines.
Only two lines have been removed (`#pramga omp end critical`); the rest are the same, but without the trailing whitespace.

Since OpenMP defines the scope of the `#pragma omp critical` command to be a structured_block, which is either a single statement or a brace-enclosed sequence of statements, removing the `#pragma omp end critical` directive will have no effect since the region enclosed by the directives was only a single line.