Fixing a changing order, coloring does not work bug.
Details, previously, Coloring functions (i.e. DistanceOneColoring, StarColoring, ...) will firstly check a flag indicating the graphs have been colored by this method or not. Coloring function only colors the graph if this graph haven't been colored by this method yet.

The problem is, if I just changed the ordering of the graph, and want to recolor the graph, this mechanism will prevent me get the correct coloring result.

I removed the code of checking this flag.